api/requirements.txt
# API Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
supabase==2.0.0
python-multipart==0.0.6
python-dotenv==1.0.0
asyncpg==0.29.0
alembic==1.13.0

# Existing dependencies from main project
sentence-transformers==2.2.2
scikit-learn==1.3.2
edge-tts==6.1.9
pydub==0.25.1

# PDF processing (for file upload)
PyPDF2==3.0.1
python-magic==0.4.27
